sentinel_cli/requirements.txt
rich
litellm
python-dotenv
requests
fpdf2
flask
